- Fix situation when uploadFrom() fails and thus $file is not set
git-svn-id: svn://svn.xp-framework.net/xp/trunk@12770 d2cacbed-c0f6-0310-851a-9dad52fc623a

+++ ports/classes/net/xp_framework/unittest/peer/ftp/IntegrationTest.class.php
@@ -317,7 +317,7 @@
       } catch (Throwable $e) {
         
         // Unfortunately, try { } finally does not exist...        
-        $file->delete();
+        $file && $file->delete();
         throw $e;
       }
     }
